Changes to use latest FrSkySportTelemetry library and polling
Includes the latest FrSkySportTelemetry.zip library (also containing
some custom code additions for this project), and also changes to the
code to allow it to work with this latest library. Also includes a
FrSky telemetry polling option which is required if not directly
connecting the teensy to a FrSky X series Rx - to allow telemetry to be
polled when connecting the teensy directly to the Taranis s.port pin in
the module bay (such as when using a ULRS Tx module/board).