v1.7rc-1
This commit includes the option for using mavlink RSSI instead of FrSky
RSSI. enabling this by uncommenting line 138: #define USE_MAV_RSSI.
This sends mavlink rxRSSI over the A3 analog sensor (and 0 over A4)
instead of the traditional roll/pitch values as is the norm. The LUA
scripts then automatically (upon detection of A4=0) displays this
mavlink RSSI in the top bar instead of the FrSky RSSI.
This commit hopefully concludes all the changes required to have this
project function for the ULRS RC system, with the teensy being ground
based, its mavlink feed coming from the UART on the ULRS Tx. Polling
mode is now available to facilitate the direct connection of the Teensy
to the Taranis radio. By default, the addons section is configured to
disable all the adding - meaning this will work in regular FrSky mode
as per the traditional design. The addons will need to be configured
for operation with ULRS - the Wiki will be updated to provide
configuration examples for this.