Removed use of model G Vars
Change #1:
The offset.lua was using the OpenTx model Global Variables table to
store selected values. This has been now removed as this is not
necessary in OpenTx - the selected values in a mixer script are held
internally by the model config itself and always accessible for any
selected model (I.e. once set, they stick). In order to get access to
these variables from the telemetry scripts, offset.lua now writes these
values to LUA global variables. The two main telemetry scripts have
been updated to now read the required LUA global variables instead of
reading from the OpenTx model global variables table.
Change #2:
Have added new function checkGlobals() to telemetry scripts - this is a
protection against a user failing to install the offset.lua, and will
initialise the global variables should offset.lua not be there to do
this. It will essenntially control the default values otherwise
provided by offset.lua, so if preferred, a user can set the values here
and do away with offset.lua.
Change #3:
The BatCap Wh input within offset.lua has had its default value changed
to 123. This is not necessarily the preferred default, but is necessary
to allow the selection of the entire range 0-250. This is due to an
OpenTx limitation as the underlying variable which holds these input
values is only 8 bits in size. OpenTx limits input values to
-128<default>127, so by setting default to 123 we can go all the way
from 0 to 250 (before with default set at 30 it was only possible to
select 0-157)
Change #4:
Have removed /SCRIPTS/Model_Name/telem1.lua as this is for OpenTx 2.0
and this is the 2.1 repo.